head

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: sivinska <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 17:30:29 by sivinska          #+#    #+#             */
-/*   Updated: 2018/04/25 15:27:25 by sivinska         ###   ########.fr       */
+/*   Updated: 2018/04/25 15:41:14 by sivinska         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -89,7 +89,7 @@ char				*ft_convert_to_base(unsigned long long nbr, int base);
 char				*ft_convert_to_base_long(unsigned long int nbr, int base);
 char				*ft_str_to_upper(char *str);
 char				*ft_strrev(char *str);
-int					ft_get_nbr_len(int n);
+int					ft_get_nbr_len(long long int n);
 void				ft_putnbr_unsigned(unsigned int nbr);
 int					ft_putchar_ret(char c);
 void				ft_putchar_unicode(unsigned int c);
